feat: search flight, create booking

+++ backend/src/main/java/com/flight/backend/dto/flight/CreateFlightRequest.java
@@ -1,5 +1,6 @@
 package com.flight.backend.dto.flight;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.flight.backend.entity.enums.FlightStatus;
@@ -15,6 +16,6 @@ public class CreateFlightRequest {
     private Long routeId;
     private LocalDateTime departureTime;
     private int estimateDuration;
-    private Long basePrice;
+    private BigDecimal basePrice;
     private FlightStatus status;
 }
