+++ backend/src/main/java/com/flight/backend/dto/flight/FlightResponse.java
@@ -1,5 +1,6 @@
 package com.flight.backend.dto.flight;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.flight.backend.entity.enums.FlightStatus;
@@ -20,7 +21,7 @@ public class FlightResponse {
     private String flightCode;
     private LocalDateTime departureTime;
     private int estimateDuration;
-    private Long basePrice;
+    private BigDecimal basePrice;
     private FlightStatus status;
     private Long airplaneId;
     private Long routeId;
