+++ backend/src/main/java/com/flight/backend/dto/flight/FlightSearchResponse.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -17,27 +18,27 @@ import java.time.LocalDateTime;
 public class FlightSearchResponse {
     private Long id;
     private String flightCode;
-    
+
     // Airline details
     private Long airlineId;
     private String airlineName;
     private String airlineLogo;
-    
+
     // Airport details
     private Long departureAirportId;
     private String departureAirportCode;
     private String departureCity;
-    
+
     private Long arrivalAirportId;
     private String arrivalAirportCode;
     private String arrivalCity;
-    
+
     // Times
     private LocalDateTime departureTime;
     private LocalDateTime arrivalTime;
     private int estimateDuration;
-    
+
     // Price
-    private Long basePrice;
+    private BigDecimal basePrice;
     private FlightStatus status;
 }
